simple output like %c %s works. %x breaks. makefile and header must be better

+++ printf/convert/x_convert.c
@@ -6,11 +6,11 @@
 /*   By: dbrittan <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/25 11:54:11 by dbrittan          #+#    #+#             */
-/*   Updated: 2020/11/25 17:11:56 by dbrittan         ###   ########.fr       */
+/*   Updated: 2020/11/26 14:56:41 by dbrittan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "../libft/libft.h"
 #include <stdio.h>
 
 static char	convert_hex(int n, char flag)
@@ -23,7 +23,7 @@ static char	convert_hex(int n, char flag)
 	return (0);
 }
 
-char	*x_convert(unsigned long int val, char flag)
+char	*x_convert(int val, char flag)
 {
 	int orig_val;
 	int len;
